+++ app/src/main/AndroidManifest.xml
@@ -45,7 +45,7 @@
         </receiver>
         <service android:name="com.google.android.gms.analytics.CampaignTrackingService" />
         <service
-            android:name=".MyFirebaseMessagingService">
+            android:name=".firebase.MyFirebaseMessagingService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT"/>
             </intent-filter>
