+++ app/src/main/java/com/tytanapps/ptsd/facility/Facility.java
@@ -1,4 +1,4 @@
-package com.tytanapps.ptsd;
+package com.tytanapps.ptsd.facility;
 
 import android.graphics.Bitmap;
 import android.support.annotation.NonNull;
