+++ app/src/main/res/layout/fragment_news.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="com.tytanapps.ptsd.fragments.NewsFragment">
+    tools:context="com.tytanapps.ptsd.news.NewsFragment">
 
     <com.bhargavms.dotloader.DotLoader
         android:id="@+id/news_progressbar"
